The gameboard its self is a reusable component.  when play again is clicked now doesnt render animation.
